# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ScopeWise AI API",
        "description": "REST API for software requirement analysis, scope-creep prevention and project estimation. Authenticate with a Sanctum bearer token obtained from POST /api/auth/login.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{scheme}://{host}",
            "description": "Current environment",
            "variables": {
                "scheme": {
                    "enum": [
                        "http",
                        "https"
                    ],
                    "default": "http"
                },
                "host": {
                    "default": "localhost:8000"
                }
            }
        }
    ],
    "paths": {
...
        "/api/dashboard/stats": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Aggregate statistics for the current user",
                "operationId": "654e1617f74d2b9e0654ec77620d960b",
                "responses": {
                    "200": {
                        "description": "Project and analysis totals, status counts"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/activity": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Recent activity for the current user",
                "operationId": "2199928c7bef307a6ce7f20f4bbb6435",
                "responses": {
                    "200": {
                        "description": "Recent projects and analysed requirements"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/forgot-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Request a password reset link",
                "description": "Always answers 200 when the payload is valid, regardless of whether the account exists, to avoid leaking addresses.",
                "operationId": "2bc87cdee42a3e7a0332d86054875d5b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jane@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reset link sent (or deliberately undisclosed)"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "429": {
                        "description": "Too many requests"
                    }
                }
            }
        },
        "/api/auth/reset/{token}": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Validate a password reset token",
                "description": "The link in the email points here with ?email= for a direct lookup; without it the most recent tokens are scanned.",
                "operationId": "91dde6cb3d553cafc9b74050024887da",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The address the link was sent to",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "email"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token valid; email echoed for form prefill"
                    },
                    "404": {
                        "description": "Invalid, expired or unknown token"
                    },
                    "429": {
                        "description": "Too many requests"
                    }
                }
            }
        },
        "/api/auth/reset": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Reset the password with the emailed token",
                "description": "On success every previously issued api token is revoked.",
                "operationId": "2392295c21c1f4ecf4a4babd5f68bc13",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset"
                    },
                    "404": {
                        "description": "Invalid or expired token"
                    },
                    "422": {
                        "description": "Validation failed or token/email mismatch"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Sanctum token issued by POST /api/auth/login. Send the token as: Authorization: Bearer <token>"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}